+++ relay-cardinality/src/redis/limiter.rs
@@ -4,16 +4,14 @@ use relay_redis::{Connection, RedisPool};
 use relay_statsd::metric;
 
 use crate::{
-    limiter::{Entry, EntryId, Limiter, Rejections, Scoping},
+    limiter::{Entry, Limiter, Rejections, Scoping},
     redis::{
         cache::{Cache, CacheOutcome},
         quota::QuotaScoping,
         script::{CardinalityScript, Status},
+        state::{LimitState, RedisEntry},
     },
-    statsd::{
-        CardinalityLimiterCounters, CardinalityLimiterHistograms, CardinalityLimiterSets,
-        CardinalityLimiterTimers,
-    },
+    statsd::{CardinalityLimiterHistograms, CardinalityLimiterTimers},
     CardinalityLimit, Result,
 };
 use relay_common::time::UnixTimestamp;
@@ -58,12 +56,12 @@ impl RedisSetLimiter {
     fn check_limits(
         &self,
         con: &mut Connection,
-        state: &mut LimitState<'_>,
+        state: &LimitState<'_>,
+        scope: QuotaScoping,
+        entries: Vec<RedisEntry>,
         timestamp: UnixTimestamp,
     ) -> Result<CheckedLimits> {
-        let scope = state.scope;
         let limit = state.limit;
-        let entries = state.take_entries();
 
         metric!(
             histogram(CardinalityLimiterHistograms::RedisCheckHashes) = entries.len() as u64,
@@ -94,7 +92,7 @@ impl RedisSetLimiter {
 }
 
 impl Limiter for RedisSetLimiter {
-    fn check_cardinality_limits<'a, E, R>(
+    fn check_cardinality_limits<'a, 'b, E, R>(
         &self,
         scoping: Scoping,
         limits: &'a [CardinalityLimit],
@@ -102,7 +100,7 @@ impl Limiter for RedisSetLimiter {
         rejections: &mut R,
     ) -> Result<()>
     where
-        E: IntoIterator<Item = Entry>,
+        E: IntoIterator<Item = Entry<'b>>,
         R: Rejections<'a>,
     {
         #[cfg(not(test))]
@@ -116,12 +114,12 @@ impl Limiter for RedisSetLimiter {
         let cache = self.cache.read(timestamp); // Acquire a read lock.
         for entry in entries {
             for state in states.iter_mut() {
-                if !state.scope.matches(&entry) {
+                let Some(scope) = state.matches(entry) else {
                     // Entry not relevant for limit.
                     continue;
-                }
+                };
 
-                match cache.check(state.scope, entry.hash, state.limit) {
+                match cache.check(scope, entry.hash, state.limit) {
                     CacheOutcome::Accepted => {
                         // Accepted already, nothing to do.
                         state.cache_hit();
@@ -135,7 +133,7 @@ impl Limiter for RedisSetLimiter {
                     }
                     CacheOutcome::Unknown => {
                         // Add the entry to the state -> needs to be checked with Redis.
-                        state.entries.push(RedisEntry::new(entry.id, entry.hash));
+                        state.add(scope, RedisEntry::new(entry.id, entry.hash));
                         state.cache_miss();
                     }
                 }
@@ -147,161 +145,29 @@ impl Limiter for RedisSetLimiter {
         let mut connection = client.connection()?;
 
         for mut state in states {
-            if state.entries.is_empty() {
-                continue;
-            }
+            for (scope, entries) in state.take_scopes() {
+                let results = metric!(timer(CardinalityLimiterTimers::Redis), id = state.id(), {
+                    self.check_limits(&mut connection, &state, scope, entries, timestamp)
+                })?;
 
-            let results = metric!(timer(CardinalityLimiterTimers::Redis), id = state.id(), {
-                self.check_limits(&mut connection, &mut state, timestamp)
-            })?;
-
-            // This always acquires a write lock, but we only hit this
-            // if we previously didn't satisfy the request from the cache,
-            // -> there is a very high chance we actually need the lock.
-            let mut cache = self.cache.update(state.scope, timestamp); // Acquire a write lock.
-            for (entry, status) in results {
-                if status.is_rejected() {
-                    rejections.reject(state.cardinality_limit(), entry.id);
-                    state.rejected();
-                } else {
-                    cache.accept(entry.hash);
-                    state.accepted();
+                // This always acquires a write lock, but we only hit this
+                // if we previously didn't satisfy the request from the cache,
+                // -> there is a very high chance we actually need the lock.
+                let mut cache = self.cache.update(scope, timestamp); // Acquire a write lock.
+                for (entry, status) in results {
+                    if status.is_rejected() {
+                        rejections.reject(state.cardinality_limit(), entry.id);
+                        state.rejected();
+                    } else {
+                        cache.accept(entry.hash);
+                        state.accepted();
+                    }
                 }
+                drop(cache); // Give up the cache lock!
             }
-            drop(cache); // Give up the cache lock!
         }
 
         Ok(())
-    }
-}
-
-/// Internal state combining relevant entries for the respective quota.
-struct LimitState<'a> {
-    /// Entries which are relevant for the quota.
-    pub entries: Vec<RedisEntry>,
-    /// Scoping of the quota.
-    pub scope: QuotaScoping,
-    /// The limit of the quota.
-    pub limit: u64,
-
-    /// The original cardinality limit.
-    cardinality_limit: &'a CardinalityLimit,
-    /// The original/full scoping.
-    scoping: Scoping,
-    /// Amount of cache hits.
-    cache_hits: i64,
-    /// Amount of cache misses.
-    cache_misses: i64,
-    /// Amount of accepts,
-    accepts: i64,
-    /// Amount of rejections.
-    rejections: i64,
-}
-
-impl<'a> LimitState<'a> {
-    pub fn new(scoping: Scoping, cardinality_limit: &'a CardinalityLimit) -> Option<Self> {
-        Some(Self {
-            entries: Vec::new(),
-            scope: QuotaScoping::new(scoping, cardinality_limit)?,
-            limit: cardinality_limit.limit,
-            cardinality_limit,
-            scoping,
-            cache_hits: 0,
-            cache_misses: 0,
-            accepts: 0,
-            rejections: 0,
-        })
-    }
-
-    pub fn from_limits(scoping: Scoping, limits: &'a [CardinalityLimit]) -> Vec<Self> {
-        limits
-            .iter()
-            .filter_map(|limit| LimitState::new(scoping, limit))
-            .collect::<Vec<_>>()
-    }
-
-    pub fn id(&self) -> &'a str {
-        &self.cardinality_limit.id
-    }
-
-    pub fn cardinality_limit(&self) -> &'a CardinalityLimit {
-        self.cardinality_limit
-    }
-
-    pub fn take_entries(&mut self) -> Vec<RedisEntry> {
-        std::mem::take(&mut self.entries)
-    }
-
-    pub fn cache_hit(&mut self) {
-        self.cache_hits += 1;
-    }
-
-    pub fn cache_miss(&mut self) {
-        self.cache_misses += 1;
-    }
-
-    pub fn accepted(&mut self) {
-        self.accepts += 1;
-    }
-
-    pub fn rejected(&mut self) {
-        self.rejections += 1;
-    }
-}
-
-impl<'a> Drop for LimitState<'a> {
-    fn drop(&mut self) {
-        let passive = if self.cardinality_limit.passive {
-            "true"
-        } else {
-            "false"
-        };
-
-        metric!(
-            counter(CardinalityLimiterCounters::RedisCacheHit) += self.cache_hits,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::RedisCacheMiss) += self.cache_misses,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::Accepted) += self.accepts,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::Rejected) += self.rejections,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-
-        let organization_id = self.scoping.organization_id as i64;
-        let status = if self.rejections > 0 { "limited" } else { "ok" };
-        metric!(
-            set(CardinalityLimiterSets::Organizations) = organization_id,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-            status = status,
-        )
-    }
-}
-
-/// Entry used by the Redis limiter.
-#[derive(Clone, Copy, Debug)]
-struct RedisEntry {
-    /// The correlating entry id.
-    pub id: EntryId,
-    /// The entry hash.
-    pub hash: u32,
-}
-
-impl RedisEntry {
-    /// Creates a new Redis entry.
-    fn new(id: EntryId, hash: u32) -> Self {
-        Self { id, hash }
     }
 }
 
@@ -329,10 +195,11 @@ mod tests {
     use std::collections::HashSet;
     use std::sync::atomic::AtomicU64;
 
-    use relay_base_schema::metrics::MetricNamespace;
+    use relay_base_schema::metrics::MetricNamespace::*;
     use relay_base_schema::project::ProjectId;
     use relay_redis::{redis, RedisConfigOptions};
 
+    use crate::limiter::EntryId;
     use crate::redis::{KEY_PREFIX, KEY_VERSION};
     use crate::{CardinalityScope, SlidingWindow};
 
@@ -371,6 +238,12 @@ mod tests {
 
     #[derive(Debug, Default, PartialEq, Eq)]
     struct Rejections(HashSet<EntryId>);
+
+    impl Rejections {
+        fn contains_any(&self, ids: impl IntoIterator<Item = usize>) -> bool {
+            ids.into_iter().any(|id| self.0.contains(&EntryId(id)))
+        }
+    }
 
     impl<'a> super::Rejections<'a> for Rejections {
         fn reject(&mut self, _limit: &'a CardinalityLimit, entry_id: EntryId) {
@@ -436,14 +309,14 @@ mod tests {
                 .collect()
         }
 
-        fn test_limits<I>(
+        fn test_limits<'a, I>(
             &self,
             scoping: Scoping,
-            limits: &[CardinalityLimit],
+            limits: &'a [CardinalityLimit],
             entries: I,
         ) -> Rejections
         where
-            I: IntoIterator<Item = Entry>,
+            I: IntoIterator<Item = Entry<'a>>,
         {
             let mut outcomes = Rejections::default();
             self.check_cardinality_limits(scoping, limits, entries, &mut outcomes)
@@ -457,12 +330,12 @@ mod tests {
         let limiter = build_limiter();
 
         let entries = [
-            Entry::new(EntryId(0), MetricNamespace::Custom, 0),
-            Entry::new(EntryId(1), MetricNamespace::Custom, 1),
-            Entry::new(EntryId(2), MetricNamespace::Custom, 2),
-            Entry::new(EntryId(3), MetricNamespace::Custom, 3),
-            Entry::new(EntryId(4), MetricNamespace::Custom, 4),
-            Entry::new(EntryId(5), MetricNamespace::Custom, 5),
+            Entry::new(EntryId(0), Custom, "a", 0),
+            Entry::new(EntryId(1), Custom, "b", 1),
+            Entry::new(EntryId(2), Custom, "c", 2),
+            Entry::new(EntryId(3), Custom, "d", 3),
+            Entry::new(EntryId(4), Custom, "e", 4),
+            Entry::new(EntryId(5), Custom, "f", 5),
         ];
 
         let scoping = new_scoping(&limiter);
@@ -475,7 +348,7 @@ mod tests {
             },
             limit: 5,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         };
 
         // 6 items, limit is 5 -> 1 rejection.
@@ -492,6 +365,38 @@ mod tests {
         limit.limit = 6;
         let rejected3 = limiter.test_limits(scoping, &[limit], entries);
         assert_eq!(rejected3.len(), 0);
+    }
+
+    #[test]
+    fn test_limiter_name_limit() {
+        let limiter = build_limiter();
+
+        let entries = [
+            Entry::new(EntryId(0), Custom, "a", 0),
+            Entry::new(EntryId(1), Custom, "a", 1),
+            Entry::new(EntryId(2), Custom, "a", 2),
+            Entry::new(EntryId(3), Custom, "b", 3),
+            Entry::new(EntryId(4), Custom, "b", 4),
+            Entry::new(EntryId(5), Custom, "b", 5),
+        ];
+
+        let scoping = new_scoping(&limiter);
+        let limit = CardinalityLimit {
+            id: "limit".to_owned(),
+            passive: false,
+            window: SlidingWindow {
+                window_seconds: 3600,
+                granularity_seconds: 360,
+            },
+            limit: 2,
+            scope: CardinalityScope::Name,
+            namespace: Some(Custom),
+        };
+
+        let rejected = limiter.test_limits(scoping, &[limit.clone()], entries);
+        assert_eq!(rejected.len(), 2);
+        assert!(rejected.contains_any([0, 1, 2]));
+        assert!(rejected.contains_any([3, 4, 5]));
     }
 
     #[test]
@@ -519,15 +424,15 @@ mod tests {
             },
             limit: 1,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
-        let entries1 = [Entry::new(EntryId(0), MetricNamespace::Custom, 0)];
+        let entries1 = [Entry::new(EntryId(0), Custom, "a", 0)];
         assert!(limiter.test_limits(scoping1, limits, entries1).is_empty());
         assert!(limiter.test_limits(scoping2, limits, entries1).is_empty());
 
         // Make sure `entries2` is not accepted.
-        let entries2 = [Entry::new(EntryId(1), MetricNamespace::Custom, 1)];
+        let entries2 = [Entry::new(EntryId(1), Custom, "a", 1)];
         assert_eq!(limiter.test_limits(scoping1, limits, entries2).len(), 1);
         assert_eq!(limiter.test_limits(scoping2, limits, entries2).len(), 1);
 
@@ -580,17 +485,17 @@ mod tests {
             },
             limit: 10_000,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
         let entries = (0..50)
-            .map(|i| Entry::new(EntryId(i as usize), MetricNamespace::Custom, i))
+            .map(|i| Entry::new(EntryId(i as usize), Custom, "a", i))
             .collect::<Vec<_>>();
         let rejected = limiter.test_limits(scoping, limits, entries);
         assert_eq!(rejected.len(), 0);
 
         let entries = (100..150)
-            .map(|i| Entry::new(EntryId(i as usize), MetricNamespace::Custom, i))
+            .map(|i| Entry::new(EntryId(i as usize), Custom, "a", i))
             .collect::<Vec<_>>();
         let rejected = limiter.test_limits(scoping, limits, entries);
         assert_eq!(rejected.len(), 0);
@@ -610,11 +515,11 @@ mod tests {
             },
             limit: 80_000,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
         let entries = (0..100_000)
-            .map(|i| Entry::new(EntryId(i as usize), MetricNamespace::Custom, i))
+            .map(|i| Entry::new(EntryId(i as usize), Custom, "a", i))
             .collect::<Vec<_>>();
 
         let rejected = limiter.test_limits(scoping, limits, entries);
@@ -636,11 +541,11 @@ mod tests {
             window,
             limit: 1,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
-        let entries1 = [Entry::new(EntryId(0), MetricNamespace::Custom, 0)];
-        let entries2 = [Entry::new(EntryId(1), MetricNamespace::Custom, 1)];
+        let entries1 = [Entry::new(EntryId(0), Custom, "a", 0)];
+        let entries2 = [Entry::new(EntryId(1), Custom, "b", 1)];
 
         // 1 item and limit is 1 -> No rejections.
         let rejected = limiter.test_limits(scoping, limits, entries1);
@@ -685,9 +590,9 @@ mod tests {
             namespace: None,
         }];
 
-        let entries1 = [Entry::new(EntryId(0), MetricNamespace::Custom, 0)];
-        let entries2 = [Entry::new(EntryId(0), MetricNamespace::Spans, 1)];
-        let entries3 = [Entry::new(EntryId(0), MetricNamespace::Transactions, 2)];
+        let entries1 = [Entry::new(EntryId(0), Custom, "a", 0)];
+        let entries2 = [Entry::new(EntryId(0), Spans, "b", 1)];
+        let entries3 = [Entry::new(EntryId(0), Transactions, "c", 2)];
 
         let rejected = limiter.test_limits(scoping, limits, entries1);
         assert_eq!(rejected.len(), 0);
@@ -714,7 +619,7 @@ mod tests {
                 },
                 limit: 1,
                 scope: CardinalityScope::Organization,
-                namespace: Some(MetricNamespace::Custom),
+                namespace: Some(Custom),
             },
             CardinalityLimit {
                 id: "limit2".to_owned(),
@@ -725,7 +630,7 @@ mod tests {
                 },
                 limit: 1,
                 scope: CardinalityScope::Organization,
-                namespace: Some(MetricNamespace::Custom),
+                namespace: Some(Custom),
             },
             CardinalityLimit {
                 id: "limit3".to_owned(),
@@ -736,7 +641,7 @@ mod tests {
                 },
                 limit: 1,
                 scope: CardinalityScope::Organization,
-                namespace: Some(MetricNamespace::Spans),
+                namespace: Some(Spans),
             },
             CardinalityLimit {
                 id: "unknown_skipped".to_owned(),
@@ -747,17 +652,17 @@ mod tests {
                 },
                 limit: 1,
                 scope: CardinalityScope::Unknown,
-                namespace: Some(MetricNamespace::Transactions),
+                namespace: Some(Transactions),
             },
         ];
 
         let entries = [
-            Entry::new(EntryId(0), MetricNamespace::Custom, 0),
-            Entry::new(EntryId(1), MetricNamespace::Custom, 1),
-            Entry::new(EntryId(2), MetricNamespace::Spans, 2),
-            Entry::new(EntryId(3), MetricNamespace::Spans, 3),
-            Entry::new(EntryId(4), MetricNamespace::Transactions, 4),
-            Entry::new(EntryId(5), MetricNamespace::Transactions, 5),
+            Entry::new(EntryId(0), Custom, "a", 0),
+            Entry::new(EntryId(1), Custom, "b", 1),
+            Entry::new(EntryId(2), Spans, "c", 2),
+            Entry::new(EntryId(3), Spans, "d", 3),
+            Entry::new(EntryId(4), Transactions, "e", 4),
+            Entry::new(EntryId(5), Transactions, "f", 5),
         ];
 
         // Run multiple times to make sure caching does not interfere.
@@ -767,9 +672,9 @@ mod tests {
             // 2 transactions + 1 span + 1 custom (4) accepted -> 2 (6-4) rejected.
             assert_eq!(rejected.len(), 2);
             // 1 custom rejected.
-            assert!(rejected.contains(&EntryId(0)) || rejected.contains(&EntryId(1)));
+            assert!(rejected.contains_any([0, 1]));
             // 1 span rejected.
-            assert!(rejected.contains(&EntryId(2)) || rejected.contains(&EntryId(3)));
+            assert!(rejected.contains_any([2, 3]));
             // 2 transactions accepted -> no rejections.
             assert!(!rejected.contains(&EntryId(4)));
             assert!(!rejected.contains(&EntryId(5)));
@@ -798,8 +703,8 @@ mod tests {
             namespace: None,
         }];
 
-        let entries1 = [Entry::new(EntryId(0), MetricNamespace::Custom, 0)];
-        let entries2 = [Entry::new(EntryId(0), MetricNamespace::Custom, 1)];
+        let entries1 = [Entry::new(EntryId(0), Custom, "a", 0)];
+        let entries2 = [Entry::new(EntryId(0), Custom, "b", 1)];
 
         // Accept different entries for different scopes.
         let rejected = limiter.test_limits(scoping1, limits, entries1);
@@ -829,13 +734,13 @@ mod tests {
             window,
             limit: 100,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
         macro_rules! test {
             ($r:expr) => {{
                 let entries = $r
-                    .map(|i| Entry::new(EntryId(i as usize), MetricNamespace::Custom, i))
+                    .map(|i| Entry::new(EntryId(i as usize), Custom, "foo", i))
                     .collect::<Vec<_>>();
 
                 limiter.test_limits(scoping, limits, entries)
@@ -913,13 +818,13 @@ mod tests {
             window,
             limit: 100,
             scope: CardinalityScope::Organization,
-            namespace: Some(MetricNamespace::Custom),
+            namespace: Some(Custom),
         }];
 
         macro_rules! test {
             ($r:expr) => {{
                 let entries = $r
-                    .map(|i| Entry::new(EntryId(i as usize), MetricNamespace::Custom, i))
+                    .map(|i| Entry::new(EntryId(i as usize), Custom, "foo", i))
                     .collect::<Vec<_>>();
 
                 limiter.test_limits(scoping, limits, entries)